LPS-171250 Extract Action Card to its own component
