+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ObjectAction/tabs/ActionBuilder.tsx
@@ -13,28 +13,19 @@
  */
 
 import ClayAlert from '@clayui/alert';
-import {ClayCheckbox, ClaySelect} from '@clayui/form';
-import ClayIcon from '@clayui/icon';
-import ClayLabel from '@clayui/label';
-import {ClayTooltipProvider} from '@clayui/tooltip';
 import {
-	API,
 	Card,
-	CodeEditor,
 	CustomItem,
-	Input,
 	InputLocalized,
-	SelectWithOption,
 	SidebarCategory,
 	SingleSelect,
 	invalidateRequired,
 } from '@liferay/object-js-components-web';
 import React, {useEffect, useMemo, useState} from 'react';
 
-import PredefinedValuesTable from '../PredefinedValuesTable';
-
 import './ActionBuilder.scss';
 import {ActionError} from '../index';
+import {ActionContainer} from './ActionContainer';
 import {ConditionContainer} from './ConditionContainer';
 interface ActionBuilderProps {
 	errors: ActionError;
@@ -50,13 +41,7 @@ interface ActionBuilderProps {
 	validateExpressionURL: string;
 	values: Partial<ObjectAction>;
 }
-
-interface SelectItem {
-	label: string;
-	value: string;
-}
-
-interface WarningStates {
+export interface WarningStates {
 	mandatoryRelationships: boolean;
 	requiredFields: boolean;
 }
@@ -91,27 +76,6 @@ export default function ActionBuilder({
 		CustomItem[]
 	>(objectActionExecutors);
 
-	const [notificationTemplates, setNotificationTemplates] = useState<
-		CustomItem<number>[]
-	>([]);
-
-	const [objectsOptions, setObjectOptions] = useState<ObjectsOptionsList>([]);
-
-	const notificationTemplateLabel = useMemo(() => {
-		return notificationTemplates.find(
-			({value}) => value === values.parameters?.notificationTemplateId
-		)?.label;
-	}, [notificationTemplates, values.parameters]);
-
-	const [relationships, setRelationships] = useState<
-		ObjectDefinitionsRelationship[]
-	>([]);
-
-	const [
-		currentObjectDefinitionFields,
-		setCurrentObjectDefinitionFields,
-	] = useState<ObjectField[]>([]);
-
 	const [infoAlert, setInfoAlert] = useState(true);
 
 	const [warningAlerts, setWarningAlerts] = useState<WarningStates>({
@@ -119,57 +83,12 @@ export default function ActionBuilder({
 		requiredFields: false,
 	});
 
+	const [
+		currentObjectDefinitionFields,
+		setCurrentObjectDefinitionFields,
+	] = useState<ObjectField[]>([]);
+
 	const [errorAlert, setErrorAlert] = useState(false);
-
-	const fetchObjectDefinitions = async () => {
-		const relationships = await API.fetchJSON<
-			ObjectDefinitionsRelationship[]
-		>(objectDefinitionsRelationshipsURL);
-
-		const relatedObjects: SelectItem[] = [];
-		const unrelatedObjects: SelectItem[] = [];
-
-		relationships?.forEach((object) => {
-			const {externalReferenceCode, id, label} = object;
-
-			const target = object.related ? relatedObjects : unrelatedObjects;
-
-			target.push({label, value: `${externalReferenceCode},${id}`});
-		});
-
-		const objectsOptionsList = [];
-
-		if (!values.parameters?.objectDefinitionExternalReferenceCode) {
-			objectsOptionsList.push({
-				disabled: true,
-				label: Liferay.Language.get('choose-an-object'),
-				selected: true,
-				value: '',
-			});
-		}
-		const fillSelect = (label: string, options: SelectItem[]) => {
-			if (options.length) {
-				objectsOptionsList.push({label, options, type: 'group'});
-			}
-		};
-
-		fillSelect(Liferay.Language.get('related-objects'), relatedObjects);
-
-		fillSelect(Liferay.Language.get('unrelated-objects'), unrelatedObjects);
-
-		setObjectOptions(objectsOptionsList);
-		setRelationships(relationships);
-	};
-
-	const actionExecutors = useMemo(() => {
-		const executors = new Map<string, string>();
-
-		newObjectActionExecutors.forEach(({label, value}) => {
-			value && executors.set(value, label);
-		});
-
-		return executors;
-	}, [newObjectActionExecutors]);
 
 	const actionTriggers = useMemo(() => {
 		const triggers = new Map<string, string>();
@@ -194,6 +113,36 @@ export default function ActionBuilder({
 	const showConditionContainer = values.objectActionTriggerKey
 		? triggerKeys.includes(values.objectActionTriggerKey)
 		: true;
+		
+	const closeWarningAlert = (warning: string) => {
+		setWarningAlerts((previousWarnings) => ({
+			...previousWarnings,
+			[warning]: false,
+		}));
+	};
+
+	useEffect(() => {
+		const predefinedValues = values.parameters?.predefinedValues;
+
+		const requiredFields = predefinedValues
+			? predefinedValues.filter(
+					({name}) => objectFieldsMap.get(name)?.required
+			  )
+			: [];
+
+		const hasEmptyValues = requiredFields?.some((item) =>
+			invalidateRequired(item.value)
+		);
+
+		setWarningAlerts((previousWarnings) => ({
+			...previousWarnings,
+			requiredFields: hasEmptyValues,
+		}));
+	}, [
+		values.parameters?.predefinedValues,
+		objectFieldsMap,
+		setWarningAlerts,
+	]);
 
 	const closeWarningAlert = (warning: string) => {
 		setWarningAlerts((previousWarnings) => ({
@@ -269,232 +218,6 @@ export default function ActionBuilder({
 		}
 		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [values.objectActionTriggerKey]);
-
-	useEffect(() => {
-		if (values.objectActionExecutorKey === 'notification') {
-			const makeFetch = async () => {
-				const NotificationTemplatesResponse = await API.getNotificationTemplates();
-
-				let notificationArray: NotificationTemplate[] = NotificationTemplatesResponse;
-
-				if (systemObject) {
-					notificationArray = NotificationTemplatesResponse.filter(
-						(notificationTemplate) =>
-							notificationTemplate.type !== 'userNotification'
-					);
-				}
-
-				setNotificationTemplates(
-					notificationArray.map(({id, name, type}) => ({
-						label: name,
-						type,
-						value: id,
-					}))
-				);
-			};
-
-			makeFetch();
-		}
-	}, [values, systemObject]);
-
-	const isValidField = ({
-		businessType,
-		objectFieldSettings,
-		system,
-	}: ObjectField) => {
-		const userRelationship = !!objectFieldSettings?.find(
-			({name, value}) =>
-				name === 'objectDefinition1ShortName' && value === 'User'
-		);
-
-		if (businessType === 'Relationship' && userRelationship) {
-			return true;
-		}
-
-		return (
-			businessType !== 'Aggregation' &&
-			businessType !== 'Formula' &&
-			businessType !== 'Relationship' &&
-			!system
-		);
-	};
-
-	const fetchObjectDefinitionFields = async () => {
-		let validFields: ObjectField[] = [];
-		let definitionId = objectDefinitionId;
-		let externalReferenceCode = objectDefinitionExternalReferenceCode;
-
-		if (values.objectActionExecutorKey === 'add-object-entry') {
-			definitionId = values?.parameters?.objectDefinitionId as number;
-			externalReferenceCode = values.parameters
-				?.objectDefinitionExternalReferenceCode as string;
-		}
-
-		if (externalReferenceCode) {
-			const items = await API.getObjectFieldsByExternalReferenceCode(
-				externalReferenceCode
-			);
-
-			validFields = items.filter(isValidField);
-		}
-
-		setCurrentObjectDefinitionFields(validFields);
-
-		const {
-			predefinedValues = [],
-		} = values.parameters as ObjectActionParameters;
-
-		const predefinedValuesMap = new Map<string, PredefinedValue>();
-
-		predefinedValues.forEach((field) => {
-			predefinedValuesMap.set(field.name, field);
-		});
-
-		const newPredefinedValues: PredefinedValue[] = [];
-
-		validFields.forEach(({label, name, required}) => {
-			if (predefinedValuesMap.has(name)) {
-				const field = predefinedValuesMap.get(name);
-
-				newPredefinedValues.push(field as PredefinedValue);
-			}
-			else if (
-				required &&
-				values.objectActionExecutorKey === 'add-object-entry'
-			) {
-				newPredefinedValues.push({
-					inputAsValue: false,
-					label,
-					name,
-					value: '',
-				});
-			}
-		});
-		setValues({
-			parameters: {
-				...values.parameters,
-				objectDefinitionExternalReferenceCode: externalReferenceCode,
-				objectDefinitionId: definitionId,
-				predefinedValues: newPredefinedValues,
-			},
-		});
-	};
-
-	const updateParameters = async (value: string) => {
-		const [externalReferenceCode, definitionIdValue] = value.split(',');
-
-		const definitionId = Number(definitionIdValue);
-
-		const object = relationships.find(
-			(relationship) =>
-				relationship.externalReferenceCode === externalReferenceCode
-		);
-
-		const parameters: ObjectActionParameters = {
-			objectDefinitionExternalReferenceCode: externalReferenceCode,
-			objectDefinitionId: definitionId,
-			predefinedValues: [],
-		};
-
-		if (object?.related) {
-			parameters.relatedObjectEntries = false;
-		}
-		const items = await API.getObjectFieldsByExternalReferenceCode(
-			externalReferenceCode
-		);
-
-		const validFields: ObjectField[] = [];
-
-		items.forEach((field) => {
-			if (isValidField(field)) {
-				validFields.push(field);
-
-				if (
-					field.required &&
-					values.objectActionExecutorKey === 'add-object-entry'
-				) {
-					(parameters.predefinedValues as PredefinedValue[]).push({
-						inputAsValue: false,
-						label: field.label,
-						name: field.name,
-						value: '',
-					});
-				}
-			}
-		});
-
-		setCurrentObjectDefinitionFields(validFields);
-
-		const normalizedParameters = {...values.parameters};
-
-		delete normalizedParameters.relatedObjectEntries;
-
-		setValues({
-			parameters: {
-				...normalizedParameters,
-				...(values.objectActionExecutorKey === 'add-object-entry' && {
-					...parameters,
-				}),
-			},
-		});
-
-		setWarningAlerts((previousWarnings) => ({
-			...previousWarnings,
-			mandatoryRelationships: items.some(
-				(field) =>
-					field.businessType === 'Relationship' &&
-					field.required === true
-			),
-		}));
-	};
-
-	const handleSelectObject = async ({
-		target: {value},
-	}: React.ChangeEvent<HTMLSelectElement>) => {
-		updateParameters(value);
-	};
-
-	useEffect(() => {
-		if (values.objectActionExecutorKey === 'add-object-entry') {
-			fetchObjectDefinitions();
-			fetchObjectDefinitionFields();
-		}
-		else if (values.objectActionExecutorKey === 'update-object-entry') {
-			updateParameters(objectDefinitionExternalReferenceCode);
-			fetchObjectDefinitionFields();
-		}
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [values.objectActionExecutorKey]);
-
-	useEffect(() => {
-		const predefinedValues = values.parameters?.predefinedValues;
-
-		const requiredFields = predefinedValues
-			? predefinedValues.filter(
-					({name}) => objectFieldsMap.get(name)?.required
-			  )
-			: [];
-
-		const hasEmptyValues = requiredFields?.some((item) =>
-			invalidateRequired(item.value)
-		);
-
-		setWarningAlerts((previousWarnings) => ({
-			...previousWarnings,
-			requiredFields: hasEmptyValues,
-		}));
-	}, [
-		values.parameters?.predefinedValues,
-		objectFieldsMap,
-		setWarningAlerts,
-	]);
-
-	const closeWarningAlert = (warning: string) => {
-		setWarningAlerts((previousWarnings) => ({
-			...previousWarnings,
-			[warning]: false,
-		}));
-	};
 
 	return (
 		<>
@@ -590,224 +313,29 @@ export default function ActionBuilder({
 				</ClayAlert>
 			)}
 
-			<Card title={Liferay.Language.get('action')}>
-				<Card
-					title={Liferay.Language.get('then[object]')}
-					viewMode="inline"
-				>
-					<div className="lfr-object__action-builder-then">
-						<SingleSelect
-							error={errors.objectActionExecutorKey}
-							onChange={({value}) => {
-								if (value === 'add-object-entry') {
-									fetchObjectDefinitions();
-								}
-
-								if (values.objectActionExecutorKey !== value) {
-									return setValues({
-										objectActionExecutorKey: value,
-										parameters: {},
-									});
-								}
-							}}
-							options={
-								Liferay.FeatureFlags['LPS-153714']
-									? newObjectActionExecutors
-									: objectActionExecutors
-							}
-							placeholder={Liferay.Language.get(
-								'choose-an-action'
-							)}
-							value={actionExecutors.get(
-								values.objectActionExecutorKey ?? ''
-							)}
-						/>
-
-						{values.objectActionExecutorKey ===
-							'add-object-entry' && (
-							<>
-								on
-								<SelectWithOption
-									aria-label={Liferay.Language.get(
-										'choose-an-object'
-									)}
-									error={
-										errors.objectDefinitionExternalReferenceCode
-									}
-									onChange={handleSelectObject}
-									options={objectsOptions}
-									value={
-										values.parameters
-											?.objectDefinitionExternalReferenceCode
-											? `${values.parameters.objectDefinitionExternalReferenceCode},${values.parameters.objectDefinitionId}`
-											: ''
-									}
-								/>
-								{values.parameters?.relatedObjectEntries !==
-									undefined && (
-									<>
-										<ClayCheckbox
-											checked={
-												values.parameters
-													.relatedObjectEntries ===
-												true
-											}
-											disabled={false}
-											label={Liferay.Language.get(
-												'also-relate-entries'
-											)}
-											onChange={({target: {checked}}) => {
-												setValues({
-													parameters: {
-														...values.parameters,
-														relatedObjectEntries: checked,
-													},
-												});
-											}}
-										/>
-										<ClayTooltipProvider>
-											<div
-												data-tooltip-align="top"
-												title={Liferay.Language.get(
-													'automatically-relate-object-entries-involved-in-the-action'
-												)}
-											>
-												<ClayIcon
-													className=".lfr-object__action-builder-tooltip-icon"
-													symbol="question-circle-full"
-												/>
-											</div>
-										</ClayTooltipProvider>
-									</>
-								)}
-							</>
-						)}
-
-						{values.objectActionExecutorKey === 'notification' && (
-							<SingleSelect<CustomItem<number>>
-								className="lfr-object__action-builder-notification-then"
-								error={errors.objectActionExecutorKey}
-								onChange={({value}) => {
-									setValues({
-										parameters: {
-											...values.parameters,
-											notificationTemplateId: value,
-										},
-									});
-								}}
-								options={notificationTemplates}
-								required
-								value={notificationTemplateLabel}
-							>
-								{notificationTemplates.map(
-									(option) =>
-										option.type && (
-											<ClayLabel
-												displayType={
-													option.type === 'email'
-														? 'success'
-														: 'info'
-												}
-											>
-												{option.type === 'email'
-													? Liferay.Language.get(
-															'email'
-													  )
-													: Liferay.Language.get(
-															'user-notification'
-													  )}
-											</ClayLabel>
-										)
-								)}
-							</SingleSelect>
-						)}
-					</div>
-				</Card>
-
-				{(values.objectActionExecutorKey === 'add-object-entry' ||
-					values.objectActionExecutorKey === 'update-object-entry') &&
-					values.parameters
-						?.objectDefinitionExternalReferenceCode && (
-						<PredefinedValuesTable
-							currentObjectDefinitionFields={
-								currentObjectDefinitionFields
-							}
-							disableRequiredChecked={
-								values.objectActionExecutorKey ===
-								'update-object-entry'
-							}
-							errors={
-								errors.predefinedValues as {
-									[key: string]: string;
-								}
-							}
-							objectFieldsMap={objectFieldsMap}
-							setValues={setValues}
-							title={
-								values.objectActionExecutorKey ===
-								'update-object-entry'
-									? Liferay.Language.get('values')
-									: ''
-							}
-							validateExpressionURL={validateExpressionURL}
-							values={values}
-						/>
-					)}
-
-				{values.objectActionExecutorKey === 'webhook' && (
-					<>
-						<Input
-							error={errors.url}
-							label={Liferay.Language.get('url')}
-							name="url"
-							onChange={({target: {value}}) => {
-								setValues({
-									parameters: {
-										...values.parameters,
-										url: value,
-									},
-								});
-							}}
-							required
-							value={values.parameters?.url}
-						/>
-
-						<Input
-							label={Liferay.Language.get('secret')}
-							name="secret"
-							onChange={({target: {value}}) => {
-								setValues({
-									parameters: {
-										...values.parameters,
-										secret: value,
-									},
-								});
-							}}
-							value={values.parameters?.secret}
-						/>
-					</>
-				)}
-
-				{values.objectActionExecutorKey === 'groovy' && (
-					<CodeEditor
-						error={errors.script}
-						mode="groovy"
-						onChange={(script, lineCount) =>
-							setValues({
-								parameters: {
-									...values.parameters,
-									lineCount,
-									script,
-								},
-							})
-						}
-						sidebarElements={objectActionCodeEditorElements.filter(
-							(element) => element.label === 'Fields'
-						)}
-						value={values.parameters?.script ?? ''}
-					/>
-				)}
-			</Card>
+			<ActionContainer
+				currentObjectDefinitionFields={currentObjectDefinitionFields}
+				errors={errors}
+				newObjectActionExecutors={newObjectActionExecutors}
+				objectActionCodeEditorElements={objectActionCodeEditorElements}
+				objectActionExecutors={objectActionExecutors}
+				objectDefinitionExternalReferenceCode={
+					objectDefinitionExternalReferenceCode
+				}
+				objectDefinitionId={objectDefinitionId}
+				objectDefinitionsRelationshipsURL={
+					objectDefinitionsRelationshipsURL
+				}
+				objectFieldsMap={objectFieldsMap}
+				setCurrentObjectDefinitionFields={
+					setCurrentObjectDefinitionFields
+				}
+				setValues={setValues}
+				setWarningAlerts={setWarningAlerts}
+				systemObject={systemObject}
+				validateExpressionURL={validateExpressionURL}
+				values={values}
+			/>
 
 			{Liferay.FeatureFlags['LPS-148804'] &&
 				values.objectActionTriggerKey === 'standalone' && (
